requirements.txt
Django==2.2.17
djangorestframework==3.12.2
gunicorn==20.1.0
Pillow==9.1.0
pytz==2022.1
sqlparse==0.4.2
psycopg2-binary==2.8.6